Fix attempt to access private constructor in async_mutex.
lock_sender's connect implementation was trying to call a
private constructor of the operation class. Normally this would
be fine if connect() was a method of lock_sender because
lock_sender was a friend of operation. But because connect
was implemented as a friend tag_invoke function it did not
have access - friendship is not transitive.

Resolved by making the operation constructor public.