Remove stopToken parameter from sync_wait().
Now that we have stop_when(), and also the fallback manual injection
of a stop_token using with_query_value() we don't need to have sync_wait()
concerned with stop-tokens. If the caller wants to be able to stop an
operation then they can apply an appropriate algorithm to the sender
before passing it to sync_wait().